Use module for classes in lib sub folder to profit from autoloading
