+++ app/lib/ticket_server_adapter.rb
@@ -1,6 +1,2 @@
-class TicketServerAdapter
-  def initialize(server)
-    @server = server
-    @logger = Rails.logger
-  end
+module TicketServerAdapter
 end
