+++ config/application.rb
@@ -27,7 +27,6 @@ module Frab
 
     # Custom directories with classes and modules you want to be autoloadable.
     config.autoload_paths += %W(#{config.root}/lib)
-    config.autoload_paths += %W(#{config.root}/app/lib/ticket_server_adapter)
     config.autoload_paths += %W(#{config.root}/app/inputs)
 
     # Configure the default encoding used in templates for Ruby 1.9.
